perf(library): fix 6.8s sidebar render; split published state; cache I/O off main
The dominant cost was SidebarView.sourcesSection calling itemsInSource()
~9x per body evaluation, each pass filtering all 77,837 faces while calling
standardizedFileURL on every element (419ms/pass vs 54ms for raw .path).
That is 6,775ms per sidebar render.

Worse, searchInput was @Published on FontLibrary, which all 14 views observe
via @EnvironmentObject. SwiftUI has no per-property granularity there, so
every keystroke re-evaluated that body — roughly 7s per character typed. The
180ms debounce added in Session 2 was dead code as a result: it debounced
searchQuery, but publishing searchInput had already invalidated the whole UI.

Measured on the reference library (77,837 faces / 4,136 families / 546
foundries), release build, via the new --bench harness:

  sourcesSection (~9 passes)   6775.0ms -> 0.1ms
  displayableDefaultSources    1696.4ms -> 0.0ms
  30 visible foundry rows       252.9ms -> 9.1ms
  toggleFavorite -> render      169.9ms -> 6.6ms
  cache decode (launch)        1313.7ms -> 1227.4ms, off the main actor
  cache encode (delete/move)   1016.6ms -> 977.9ms, debounced + detached
  settled phys_footprint          437MB -> 364MB

Changes:
- Source/foundry buckets cached by derivedVersion, storing indices rather
  than FontItem copies (a struct-copy map duplicates ~17MB per map here).
  Adds itemCountInSource and foundryActivation so the sidebar's 546 rows
  never materialize item arrays.
- URLs standardized once at scan time; everything downstream compares .path.
- New Services/UIState.swift: SelectionModel and PreviewSettings own the
  high-frequency state; the search field owns its text as plain @State.
- Invalidation split into invalidateDerived (items changed) and
  invalidateMembership (favorites/collections/instances).
- loadCachedLibraryOffMain at bootstrap; scheduleCacheSave/saveCacheNow for
  writes.
- FontItem shrunk: format String -> FontFormat enum, panose [Int] -> [UInt8],
  displayName stored only when it differs from the derived "Family Style".
  Decoding stays compatible with old caches, so no forced rescan.
- FontListView uses List (recycles rows) instead of ScrollView+LazyVStack.
- FontPreviewCache caches CTFontDescriptors per file+face, independent of
  size, so dragging the preview slider no longer re-parses font files from
  disk inside a view body.

Also fixes a data-loss bug: FontItem.id hashes the absolute path, so
Organize's move changed it and the next rescan silently dropped the font out
of every Favorite, Project and Palette. migrateReferences now re-points
references using PostScript name + filename + byte size.

Known regression: peak phys_footprint 549MB -> 712MB. The cache decode now
runs concurrently with SwiftUI building the UI instead of blocking it.
Replacing the 60MB JSON cache with a binary/streaming format is queued as the
next perf task in NEXT_SESSION.md.

+++ Sources/SaewooFont/Services/FontClassifier.swift
@@ -16,7 +16,7 @@ enum FontClassifier {
                          traits: CTFontSymbolicTraits,
                          weight: Double,
                          slant: Double,
-                         panose: [Int]) -> ([FontCategory], [FontMood]) {
+                         panose: [UInt8]) -> ([FontCategory], [FontMood]) {
 
         let tokens = tokenize("\(family) \(style) \(psName)")
         let categories = categorize(traits: traits, tokens: tokens, panose: panose)
@@ -33,7 +33,7 @@ enum FontClassifier {
     /// adds an additional orthogonal tag if the traits or name say so.
     private static func categorize(traits: CTFontSymbolicTraits,
                                    tokens: Set<String>,
-                                   panose: [Int]) -> [FontCategory] {
+                                   panose: [UInt8]) -> [FontCategory] {
         var result: Set<FontCategory> = []
 
         // Orthogonal: monospace is independent of shape. Add it if traits say
@@ -96,7 +96,7 @@ enum FontClassifier {
                                  tokens: Set<String>,
                                  weight: Double,
                                  slant: Double,
-                                 panose: [Int]) -> [FontMood] {
+                                 panose: [UInt8]) -> [FontMood] {
         var moods: Set<FontMood> = []
 
         if weight >= 0.5 { moods.insert(.bold) }
